% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Instrument_Equity.R
\name{equity}
\alias{equity}
\title{Profit/loss profile of Equity}
\usage{
equity(price, commission = NULL, shares = NULL, short = NULL,
  strikeRange = NULL, granularity = NULL)
}
\arguments{
\item{price}{Price of the equity.}

\item{commission}{Commission charged for purchasing the equity. Defaults to $0.}

\item{shares}{Number of shares purchased. Deefaults to 1.}

\item{short}{Set to TRUE if a short position is opened. Defaults to FALSE.}

\item{strikeRange}{Range of strike prices for which profit/loss is calculated. Defualts to 90\% to 110\%.}

\item{granularity}{Granularity of range of prices generated. Defaults to $0.01.}
}
\description{
\code{equity} returns an object of values,
with the strike price of the underlying asset,
and the corresponding profit
}
\examples{
callOption(strike = 100, premium = 4, short = TRUE)
}
