% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Instrument_PutOption.R
\name{putOption}
\alias{putOption}
\title{Profit/loss profile of Put Option}
\usage{
putOption(strike, premium, size = NULL, short = NULL, strikeRange = NULL,
  granularity = NULL)
}
\arguments{
\item{strike}{Strike price of the put option.}

\item{premium}{Premium of the put option.}

\item{size}{Contract size of the option. Defaults to 100.}

\item{short}{Set to TRUE if a short position is opened. Defaults to FALSE.}

\item{strikeRange}{Range of strike prices for which profit/loss is calculated. Defualts to 90\% to 110\%.}

\item{granularity}{Granularity of range of prices generated. Defaults to $0.01.}
}
\description{
\code{putOption} returns an object of values,
with the x coordinate denoting the strike price
of the underlying asset, and the y coordinate denoting the profit
were the underlying to be at a price level of $x.
}
\examples{
putOption(strike = 100, premium = 4, short = TRUE)
}
