% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Derivative_Put.R
\name{putOption}
\alias{putOption}
\title{Profit/loss profile of Put Option}
\usage{
putOption(strike, premium, size = 100, short = FALSE, lower = NULL,
  upper = NULL, granularity = 0.01)
}
\arguments{
\item{strike}{Strike price of the put option.}

\item{premium}{Premium of the put option.}

\item{size}{Contract size of the option. Defaults to 100.}

\item{short}{Set to TRUE if a short position is opened. Defaults to FALSE.}

\item{lower}{Lower bound of the range of strike prices. Defaults to 90\% of strike price.}

\item{upper}{Upper bound of the range of strike prices. Defaults to 110\% of strike price.}

\item{granularity}{Granularity of range of prices generated. Defaults to $0.01.}
}
\description{
\code{putOption} returns a tuple of values,
with the x coordinate denoting the strike price
of the underlying asset, and the y coordinate denoting the profit
were the underlying to be at a price level of $x.
}
\examples{
putOption(strike = 100, premium = 4, short = TRUE)
}
